+++ project/src/components/private-route/private-route.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import {Route, Redirect, RouteProps} from 'react-router-dom';
+import {AppRoute, AuthorizationStatus} from '../../const';
+
+type PrivateRouteProps = RouteProps & {
+  render: () => JSX.Element;
+  authorizationStatus: AuthorizationStatus;
+}
+
+function PrivateRoute(props: PrivateRouteProps): JSX.Element {
+  const {exact, path, render, authorizationStatus} = props;
+  return (
+    <Route
+      exact={exact}
+      path={path}
+      render={() => (
+        authorizationStatus === AuthorizationStatus.Auth
+          ? render()
+          : <Redirect to={AppRoute.Root} />
+      )}
+    />
+  );
+}
+
+export default PrivateRoute;
