+++ project/src/views/home/home.tsx
@@ -1,4 +1,5 @@
-import OfferCard from '../offer-card/offer-card';
+import Logo from '../../components/logo/logo';
+import OfferCard from '../../components/offer-card/offer-card';
 
 type HomeProps = {
   offersCount: number,
@@ -11,9 +12,7 @@ function Home({offersCount}: HomeProps): JSX.Element {
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link header__logo-link--active" href="#/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </a>
+              <Logo/>
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
